plot: bugfix on reliability calculation
- use dataQueueFillData instead of txQueueFillData
- fix the fomula to get reliability

dataQueueFillData should be used to get reliability. dataQueueFillData
indicates the number of application packets in queues. On the other
hand, txQueueFillData indicates the number of all the packets in
queues.

In principle, end-to-end PDR, "reliability", is derived by dividing
the total number of received packets at the root node by the total
number of generated packets. However, some application pakcets may
still be queued at the end of the simulation. We subtract the number
of queued application packets from the total number of generated
packets; then using the result, we divide the total number of packets
received to get the reliability value.

In addition, to my best knowledge, we don't need to care about
packet duplication, which won't happen with the simulator.

XXX: need discussion whether keeping plot_reliability() in
plotStatsVsParameter.py or not.