#183 decrease the default ETX from 9999 to 16
A big default value such as 9999 causes network churns, where a node
eventually selects one of its children as the preferred parent after
having unicast frames to it.

Using the small default value prevents such things; but this is not a
final solution. Needs further consideration. 16 is the default value
used in OpenWSN.