#185 introduce "exec_randomSeed" setting to config.json
A user can specify a static random seed (integer) value used in the
simulator with the "exec_randomSeed". Setting "random" to
"exec_randomSeed" lets the simulator pick a seed randomly.

A seed value used in a simulation run will be logged in a log line
whose log type is "simulator.random_seed" unless the log type is
filtered out.