prop: abandon phy_minRssi
"phy_minRssi" is used to decide which transmission a listener could
hear. The default value of phy_minRssi is -97, which results in PDR of
0% under the Pister Hack model.

It's more natural to use directly PDR for this purpose. Other models
or connectivity traces could have a different relation between RSSI
and PDR. That is, some could have PDR larger than 0 for RSSI of -97,
or some could have PDR of 0 for RSSI larger than -97.