#318 change the matrix initialization logic (K7)
The connectivity matrix is initilized with trace lines which are
captured during the first one hour of the measurement. This "one hour"
is kind of abitrary and may hide link quality changes within that one
hour window.

This commit changes the initialization logic to (try to) use only the
trace lines in the first iteration of the measurement.