#363 add the autonomous cell to a new parent explicitly (MSF)
The autonomous cell is added automatically when a new neighbor entry
is inserted for a new parent. But, if the parent is the same node as
the join proxy, its autonomous cell is not added when it's selected as
the preferred parent because the neighbor entry for the parent is
already there.

This commit handles that situation.